refactor(dashboard): decompose god modules (P0: CSS + command.ts)
Split components.css (7076 lines) into 17 domain-specific CSS files
and command.ts (2948 lines) into 9 focused modules under command/.

Pure structural refactoring — zero functionality changes.
Build verified with tsc + vite.
