+++ dashboard/src/components/command/index.ts
@@ -1,0 +1,231 @@
+import { html } from 'htm/preact'
+import { useEffect } from 'preact/hooks'
+import {
+  commandPlaneActionError,
+  commandPlaneError,
+  commandPlaneLoading,
+  commandPlaneSnapshot,
+  commandPlaneSurface,
+  focusCommandPlaneChainOperation,
+  refreshCommandPlaneChainSummary,
+  refreshCommandPlaneCurrentSurface,
+  refreshCommandPlaneHelp,
+  refreshCommandPlaneSwarm,
+  runCommandPlaneDispatchTick,
+  setCommandPlaneSurface,
+} from '../../command-store'
+import { refreshOperatorSnapshot } from '../../operator-store'
+import { navigate, route } from '../../router'
+import { SurfaceSemanticIntro } from '../common/semantic-layer'
+import {
+  commandSurfaceForContext,
+  workflowContextForRoute,
+} from '../../workflow-context'
+import {
+  actionDisabled,
+  CHAIN_SSE_EVENT_TYPES,
+  chainEventsUrl,
+  COMMAND_SURFACE_GROUPS,
+  COMMAND_SURFACE_META,
+  fire,
+  isCommandSurface,
+  surfaceRouteParams,
+} from './helpers'
+import { CommandEntryStrip, CommandWorkflowBanner } from './summary-hero'
+import { DetailLoadingState, SummarySurface } from './guided-panel'
+import { WarRoomSurface } from './war-room'
+import { SwarmSurface } from './swarm'
+import { ChainsSurface, OperationsSurface } from './operations'
+import { AlertsSurface, TopologySurface, TraceSurface } from './topology'
+import { ControlSurface } from './control'
+
+function SurfaceTabs() {
+  return html`
+    <div class="command-surface-tabs grouped">
+      ${COMMAND_SURFACE_GROUPS.map(group => html`
+        <div class="command-tab-group" key=${group.id}>
+          <span class="command-tab-group-label">${group.label}</span>
+          <div class="command-tab-group-items">
+            ${COMMAND_SURFACE_META
+              .filter(surface => surface.group === group.id)
+              .map(surface => html`
+                <button
+                  class="command-surface-tab ${commandPlaneSurface.value === surface.id ? 'active' : ''}"
+                  onClick=${() => {
+                    setCommandPlaneSurface(surface.id)
+                    navigate('command', surfaceRouteParams(surface.id))
+                  }}
+                >
+                  ${surface.label}
+                </button>
+              `)}
+          </div>
+        </div>
+      `)}
+    </div>
+  `
+}
+
+function SurfaceBody() {
+  if (commandPlaneSurface.value === 'warroom') {
+    return html`<${WarRoomSurface} />`
+  }
+  if (commandPlaneSurface.value === 'summary') {
+    return html`<${SummarySurface} />`
+  }
+  if (commandPlaneSurface.value === 'swarm') {
+    return html`<${SwarmSurface} />`
+  }
+  if (!commandPlaneSnapshot.value) {
+    return html`<${DetailLoadingState} />`
+  }
+  switch (commandPlaneSurface.value) {
+    case 'chains':
+      return html`<${ChainsSurface} />`
+    case 'topology':
+      return html`<${TopologySurface} />`
+    case 'alerts':
+      return html`<${AlertsSurface} />`
+    case 'trace':
+      return html`<${TraceSurface} />`
+    case 'control':
+      return html`<${ControlSurface} />`
+    case 'operations':
+    default:
+      return html`<${OperationsSurface} />`
+  }
+}
+
+export function Command() {
+  useEffect(() => {
+    void refreshCommandPlaneCurrentSurface()
+    void refreshCommandPlaneChainSummary()
+    void refreshCommandPlaneHelp()
+    void refreshCommandPlaneSwarm()
+  }, [])
+
+  useEffect(() => {
+    if (route.value.tab !== 'command') return
+    const requestedSurface = route.value.params.surface
+    const requestedOperation = route.value.params.operation
+    const workflowContext = workflowContextForRoute(route.value)
+    if (isCommandSurface(requestedSurface)) {
+      setCommandPlaneSurface(requestedSurface)
+    }
+    else if (workflowContext) {
+      const suggestedSurface = commandSurfaceForContext(workflowContext)
+      if (isCommandSurface(suggestedSurface)) {
+        setCommandPlaneSurface(suggestedSurface)
+      }
+    }
+    else if (!requestedSurface) {
+      setCommandPlaneSurface('warroom')
+    }
+    if (requestedOperation) {
+      focusCommandPlaneChainOperation(requestedOperation)
+    }
+    if (requestedSurface === 'swarm' || requestedSurface === 'warroom' || commandPlaneSurface.value === 'warroom') {
+      void refreshCommandPlaneSwarm()
+    }
+    if (requestedSurface === 'warroom' || commandPlaneSurface.value === 'warroom') {
+      void refreshOperatorSnapshot()
+    }
+  }, [
+    route.value.tab,
+    route.value.params.surface,
+    route.value.params.operation,
+    route.value.params.operation_id,
+    route.value.params.run_id,
+    route.value.params.source,
+    route.value.params.action_type,
+    route.value.params.target_type,
+    route.value.params.target_id,
+    route.value.params.focus_kind,
+  ])
+
+  useEffect(() => {
+    let refreshTimer: ReturnType<typeof window.setTimeout> | null = null
+    const scheduleRefresh = () => {
+      if (refreshTimer) return
+      refreshTimer = window.setTimeout(() => {
+        refreshTimer = null
+        void refreshCommandPlaneCurrentSurface()
+        void refreshCommandPlaneChainSummary()
+        if (commandPlaneSurface.value === 'swarm' || commandPlaneSurface.value === 'warroom') {
+          void refreshCommandPlaneSwarm()
+        }
+        if (commandPlaneSurface.value === 'warroom') {
+          void refreshOperatorSnapshot()
+        }
+      }, 250)
+    }
+
+    const es = new EventSource(chainEventsUrl())
+    const listeners = CHAIN_SSE_EVENT_TYPES.map(type => {
+      const handler = () => scheduleRefresh()
+      es.addEventListener(type, handler)
+      return { type, handler }
+    })
+    es.onerror = () => {
+      scheduleRefresh()
+    }
+
+    return () => {
+      listeners.forEach(({ type, handler }) => {
+        es.removeEventListener(type, handler)
+      })
+      es.close()
+      if (refreshTimer) {
+        window.clearTimeout(refreshTimer)
+      }
+    }
+  }, [])
+
+  return html`
+    <section class="dashboard-panel command-plane-view">
+      <div class="panel-header">
+        <div>
+          <h2>지휘면</h2>
+          <p>기본 진입은 라이브 워룸입니다. 실제 run, worker, message, trace를 먼저 보고 필요할 때만 detail surface로 내려갑니다.</p>
+        </div>
+        <div class="panel-actions">
+          <button
+            class="control-btn ghost"
+            onClick=${() => {
+              void fire(() => runCommandPlaneDispatchTick())
+            }}
+            disabled=${actionDisabled('dispatch:tick')}
+          >
+            ${actionDisabled('dispatch:tick') ? '정리 중...' : 'Tick 실행'}
+          </button>
+          <button
+            class="control-btn ghost"
+            onClick=${() => {
+              void refreshCommandPlaneCurrentSurface()
+              void refreshCommandPlaneChainSummary()
+              void refreshCommandPlaneSwarm()
+              if (commandPlaneSurface.value === 'warroom') {
+                void refreshOperatorSnapshot()
+              }
+            }}
+            disabled=${commandPlaneLoading.value}
+          >
+            ${commandPlaneLoading.value ? '새로고침 중...' : '새로고침'}
+          </button>
+        </div>
+      </div>
+
+      ${commandPlaneError.value
+        ? html`<div class="empty-state error">${commandPlaneError.value}</div>`
+        : null}
+      ${commandPlaneActionError.value
+        ? html`<div class="empty-state error">${commandPlaneActionError.value}</div>`
+        : null}
+      <${SurfaceSemanticIntro} surfaceId="command" />
+      <${CommandWorkflowBanner} />
+      ${commandPlaneSurface.value === 'warroom' ? null : html`<${CommandEntryStrip} />`}
+      <${SurfaceTabs} />
+      <${SurfaceBody} />
+    </section>
+  `
+}
