+++ dashboard/src/styles/activity.css
@@ -1,0 +1,82 @@
+/* ── Activity Panel (global slide-out) ────────────────────── */
+
+.activity-panel-toggle {
+  padding: 6px 14px;
+  background: rgba(255, 255, 255, 0.06);
+  border: 1px solid rgba(255, 255, 255, 0.1);
+  border-radius: 8px;
+  color: #bbb;
+  font-size: 13px;
+  cursor: pointer;
+  transition: all 0.15s;
+}
+.activity-panel-toggle:hover {
+  background: rgba(255, 255, 255, 0.1);
+  color: #eee;
+}
+.activity-panel-toggle.active {
+  background: rgba(74, 222, 128, 0.12);
+  border-color: rgba(74, 222, 128, 0.3);
+  color: #4ade80;
+}
+
+.activity-panel-backdrop {
+  position: fixed;
+  inset: 0;
+  background: rgba(0, 0, 0, 0.35);
+  backdrop-filter: blur(2px);
+  z-index: 3050;
+}
+
+.activity-panel {
+  position: fixed;
+  top: 0;
+  right: 0;
+  bottom: 0;
+  width: min(520px, 90vw);
+  background: var(--bg-primary, #111318);
+  border-left: 1px solid rgba(255, 255, 255, 0.08);
+  z-index: 3051;
+  display: flex;
+  flex-direction: column;
+  animation: activitySlideIn 0.2s ease;
+}
+
+@keyframes activitySlideIn {
+  from { transform: translateX(100%); }
+  to { transform: translateX(0); }
+}
+
+.activity-panel-header {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  padding: 16px 20px;
+  border-bottom: 1px solid rgba(255, 255, 255, 0.08);
+  flex-shrink: 0;
+}
+.activity-panel-header h3 {
+  margin: 0;
+  font-size: 15px;
+  color: #eee;
+}
+.activity-panel-close {
+  padding: 4px 12px;
+  background: rgba(255, 255, 255, 0.06);
+  border: 1px solid rgba(255, 255, 255, 0.1);
+  border-radius: 6px;
+  color: #aaa;
+  font-size: 12px;
+  cursor: pointer;
+}
+.activity-panel-close:hover {
+  background: rgba(255, 255, 255, 0.1);
+  color: #eee;
+}
+
+.activity-panel-body {
+  flex: 1;
+  overflow-y: auto;
+  padding: 16px 20px;
+}
+
