+++ dashboard/src/styles/agent.css
@@ -1,0 +1,73 @@
+/* ── Agent List (Overview) ─────────────────────────────────── */
+.agent-list { display: flex; flex-direction: column; gap: 8px; }
+.agent-row {
+  display: flex;
+  align-items: center;
+  gap: 10px;
+  padding: 8px 12px;
+  background: rgba(255,255,255,0.03);
+  border-radius: 8px;
+  border: 1px solid rgba(255,255,255,0.06);
+}
+.agent-row .agent-emoji { font-size: 20px; }
+.agent-row .agent-name { flex: 1; font-size: 14px; }
+.agent-activity-meta {
+  width: 100%;
+  margin-top: 4px;
+  color: var(--text-muted);
+  font-size: 11px;
+  line-height: 1.45;
+}
+
+/* ── Live Agent / Keeper Cards ─────────────────────────────── */
+.live-agent {
+  display: flex;
+  flex-direction: column;
+  gap: 6px;
+  padding: 12px;
+  background: rgba(255,255,255,0.04);
+  border-radius: 10px;
+  border: 1px solid rgba(255,255,255,0.08);
+  transition: all 0.2s;
+}
+.live-agent:hover { border-color: rgba(255,255,255,0.15); }
+.live-agent.keeper-card {
+  cursor: pointer;
+}
+.live-agent.keeper-card.selected {
+  border-color: rgba(74,222,128,0.55);
+  box-shadow: 0 0 0 1px rgba(74,222,128,0.35) inset;
+}
+
+/* ── Agents Tab Grid ───────────────────────────────────────── */
+.agents-grid {
+  display: grid;
+  grid-template-columns: repeat(auto-fill, minmax(280px, 1fr));
+  gap: 12px;
+}
+.agent-card {
+  padding: 15px;
+  background: rgba(255,255,255,0.04);
+  border-radius: 10px;
+  border: 1px solid rgba(255,255,255,0.08);
+  transition: border-color 0.2s;
+  color: inherit;
+  font: inherit;
+  text-align: left;
+  width: 100%;
+}
+.agent-card:hover { border-color: rgba(255,255,255,0.2); }
+.agent-card .agent-header {
+  display: flex;
+  align-items: center;
+  gap: 10px;
+  margin-bottom: 8px;
+}
+.agent-card .agent-emoji { font-size: 24px; }
+.agent-card .agent-name { font-size: 15px; font-weight: 600; }
+.agent-card .agent-korean { font-size: 12px; color: #888; }
+.agent-card .agent-model { font-size: 11px; color: #22d3ee; font-family: monospace; }
+.agent-card .agent-interests {
+  display: flex; flex-wrap: wrap; gap: 4px; margin-top: 8px;
+}
+
