+++ dashboard/src/styles/badge-card.css
@@ -1,0 +1,162 @@
+/* ── Status Badge ──────────────────────────────────────────── */
+.status-badge {
+  display: inline-flex;
+  align-items: center;
+  gap: 4px;
+  font-size: 11px;
+  padding: 2px 8px;
+  border-radius: 999px;
+}
+.status-badge.active { background: rgba(74,222,128,0.15); color: #4ade80; }
+.status-badge.busy { background: rgba(251,191,36,0.15); color: #fbbf24; }
+.status-badge.listening { background: rgba(56,189,248,0.15); color: #38bdf8; }
+.status-badge.idle { background: rgba(136,136,136,0.15); color: #888; }
+.status-badge.offline { background: rgba(85,85,85,0.15); color: #555; }
+.status-badge.todo { background: rgba(136,136,136,0.15); color: #888; }
+.status-badge.in-progress, .status-badge.in_progress {
+  background: rgba(251,191,36,0.15); color: #fbbf24;
+}
+.status-badge.done, .status-badge.completed {
+  background: rgba(74,222,128,0.15); color: #4ade80;
+}
+.status-badge.running { background: rgba(251,191,36,0.15); color: #fbbf24; }
+.status-badge.interrupted { background: rgba(56,189,248,0.15); color: #38bdf8; }
+.status-badge.stopped { background: rgba(148,163,184,0.15); color: #94a3b8; }
+.status-badge.error { background: rgba(251,113,133,0.15); color: #fb7185; }
+
+/* ── Card (generic wrapper) ────────────────────────────────── */
+.card {
+  background: rgba(255,255,255,0.03);
+  border-radius: 12px;
+  padding: 20px;
+  border: 1px solid rgba(255,255,255,0.08);
+}
+.card-title {
+  font-size: 14px;
+  text-transform: uppercase;
+  letter-spacing: 2px;
+  color: #4ade80;
+  margin-bottom: 15px;
+}
+.card-title-row {
+  display: flex;
+  align-items: flex-start;
+  justify-content: space-between;
+  gap: 12px;
+  margin-bottom: 15px;
+}
+.card-title-row .card-title {
+  margin-bottom: 0;
+}
+.semantic-surface-card {
+  margin-bottom: 16px;
+  padding: 16px 18px;
+  background: rgba(74,222,128,0.08);
+  border: 1px solid rgba(74,222,128,0.16);
+  border-radius: 12px;
+}
+.semantic-surface-card.compact {
+  margin-bottom: 12px;
+  padding: 12px 14px;
+}
+.semantic-surface-head {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  gap: 8px;
+  margin-bottom: 8px;
+}
+.semantic-lead {
+  margin: 0 0 12px 0;
+  color: #d9f6e3;
+  line-height: 1.55;
+}
+.semantic-inline {
+  border: 1px solid rgba(74,222,128,0.18);
+  border-radius: 10px;
+  background: rgba(255,255,255,0.03);
+}
+.semantic-inline.compact {
+  background: transparent;
+  border-color: rgba(74,222,128,0.14);
+}
+.semantic-summary {
+  cursor: pointer;
+  list-style: none;
+  padding: 6px 10px;
+  color: #9fe6b5;
+  font-size: 11px;
+  letter-spacing: 0.08em;
+  text-transform: uppercase;
+}
+.semantic-summary::-webkit-details-marker {
+  display: none;
+}
+.semantic-body {
+  padding: 0 12px 12px 12px;
+}
+.semantic-grid {
+  display: grid;
+  grid-template-columns: 110px minmax(0, 1fr);
+  gap: 8px 12px;
+  margin-bottom: 12px;
+  font-size: 12px;
+  line-height: 1.5;
+}
+.semantic-grid > span:nth-child(odd) {
+  color: #8ea9d6;
+  text-transform: uppercase;
+  letter-spacing: 0.06em;
+  font-size: 10px;
+}
+.semantic-grid.compact {
+  grid-template-columns: 88px minmax(0, 1fr);
+}
+.semantic-tag-row {
+  display: flex;
+  flex-wrap: wrap;
+  gap: 6px;
+  margin-bottom: 12px;
+}
+.semantic-tag,
+.semantic-code {
+  display: inline-flex;
+  align-items: center;
+  gap: 4px;
+  padding: 4px 8px;
+  border-radius: 999px;
+  background: rgba(255,255,255,0.06);
+  color: #c9d8f0;
+  font-size: 11px;
+}
+.semantic-code {
+  background: rgba(34,211,238,0.12);
+  color: #9be7f3;
+  font-family: monospace;
+}
+.semantic-metric-list {
+  display: flex;
+  flex-direction: column;
+  gap: 10px;
+}
+.semantic-metric-row {
+  padding: 10px 12px;
+  border-radius: 10px;
+  background: rgba(255,255,255,0.03);
+  border: 1px solid rgba(255,255,255,0.06);
+}
+.semantic-metric-row p {
+  margin: 6px 0 10px 0;
+  color: #d3dbeb;
+}
+.semantic-metric-head {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  gap: 8px;
+}
+.semantic-inline-state {
+  color: #8ea9d6;
+  font-size: 11px;
+}
+
