+++ dashboard/src/styles/board.css
@@ -1,0 +1,291 @@
+/* ── Board (Posts) ─────────────────────────────────────────── */
+.feed-health-banner {
+  margin-bottom: 12px;
+  padding: 8px 10px;
+  border-radius: 8px;
+  border: 1px solid rgba(138, 163, 211, 0.3);
+  background: rgba(71, 184, 255, 0.12);
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  gap: 10px;
+  flex-wrap: wrap;
+}
+
+.feed-health-banner.degraded {
+  border-color: rgba(251, 191, 36, 0.4);
+  background: rgba(251, 191, 36, 0.12);
+}
+
+.feed-health-title {
+  font-size: 12px;
+  font-weight: 600;
+  color: #d8e8ff;
+}
+
+.feed-health-banner.degraded .feed-health-title {
+  color: #ffe09b;
+}
+
+.feed-health-meta {
+  font-size: 11px;
+  color: #9fb8e1;
+}
+
+.feed-health-banner.degraded .feed-health-meta {
+  color: #fcd985;
+}
+
+.board-controls {
+  display: flex;
+  gap: 8px;
+  align-items: center;
+  flex-wrap: wrap;
+}
+
+.board-toolbar {
+  display: flex;
+  justify-content: space-between;
+  gap: 12px;
+  align-items: flex-start;
+  flex-wrap: wrap;
+  margin-bottom: 15px;
+}
+
+.board-toolbar-actions {
+  display: flex;
+  gap: 8px;
+  flex-wrap: wrap;
+}
+
+.board-summary-strip {
+  display: grid;
+  grid-template-columns: repeat(auto-fit, minmax(170px, 1fr));
+  gap: 10px;
+  margin-bottom: 12px;
+}
+
+.board-summary-item {
+  padding: 12px 14px;
+  border-radius: 10px;
+  border: 1px solid rgba(138, 163, 211, 0.18);
+  background: rgba(13, 22, 40, 0.62);
+  display: flex;
+  flex-direction: column;
+  gap: 4px;
+}
+
+.board-summary-label {
+  color: var(--text-muted);
+  font-size: 11px;
+  letter-spacing: 0.06em;
+  text-transform: uppercase;
+}
+
+.board-summary-item strong {
+  color: var(--text-strong);
+  font-size: 14px;
+}
+.board-sort-btn {
+  padding: 6px 14px;
+  border: 1px solid rgba(255,255,255,0.1);
+  border-radius: 8px;
+  background: rgba(255,255,255,0.04);
+  color: #888;
+  cursor: pointer;
+  font-size: 12px;
+  transition: all 0.2s;
+}
+.board-sort-btn:hover { background: rgba(255,255,255,0.08); color: #ccc; }
+.board-sort-btn.active {
+  background: rgba(74,222,128,0.15);
+  color: #4ade80;
+  border-color: rgba(74,222,128,0.3);
+}
+.board-title { font-size: 16px; font-weight: 600; margin-bottom: 12px; }
+
+.board-post {
+  display: flex;
+  gap: 10px;
+  padding: 14px 14px 13px;
+  background: rgba(255,255,255,0.04);
+  border-radius: 12px;
+  cursor: pointer;
+  border: 1px solid rgba(255,255,255,0.06);
+  margin-bottom: 10px;
+  transition: border-color 0.2s, transform 0.2s, background 0.2s;
+}
+.board-post:hover {
+  border-color: rgba(71, 184, 255, 0.26);
+  background: rgba(255,255,255,0.06);
+  transform: translateY(-1px);
+}
+
+/* Vote column */
+.vote-column {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  gap: 4px;
+  min-width: 52px;
+  padding-top: 2px;
+}
+.vote-btn {
+  background: none;
+  border: none;
+  color: #666;
+  cursor: pointer;
+  font-size: 16px;
+  padding: 2px;
+  transition: color 0.2s;
+}
+.vote-btn:hover { color: #ccc; }
+.vote-btn.upvote:hover, .vote-btn.upvote.active { color: #ff4500; }
+.vote-btn.downvote:hover, .vote-btn.downvote.active { color: #7193ff; }
+.vote-btn.animate { animation: votePop 0.3s ease; }
+.vote-count { font-size: 14px; font-weight: bold; color: #ccc; }
+
+/* Post content */
+.post-content {
+  flex: 1;
+  min-width: 0;
+  display: grid;
+  gap: 10px;
+}
+
+.post-head {
+  display: grid;
+  gap: 8px;
+}
+
+.post-title-row {
+  display: flex;
+  justify-content: space-between;
+  gap: 12px;
+  align-items: flex-start;
+}
+
+.post-title {
+  font-size: 15px;
+  font-weight: 650;
+  line-height: 1.45;
+  color: var(--text-strong);
+}
+
+.post-chip-row {
+  display: inline-flex;
+  gap: 6px;
+  align-items: center;
+  flex-wrap: wrap;
+  justify-content: flex-end;
+}
+
+.board-meta-chip {
+  font-size: 10px;
+  padding: 2px 8px;
+  border-radius: 999px;
+  border: 1px solid rgba(71, 184, 255, 0.3);
+  color: #9ad9ff;
+  background: rgba(71, 184, 255, 0.12);
+}
+
+.post-meta {
+  display: flex;
+  gap: 10px;
+  flex-wrap: wrap;
+  font-size: 11px;
+  color: #90a4cc;
+}
+
+.post-snippet {
+  color: var(--text-body);
+  font-size: 13px;
+  line-height: 1.6;
+}
+.post-author-avatar {
+  width: 24px;
+  height: 24px;
+  border-radius: 50%;
+  display: inline-flex;
+  align-items: center;
+  justify-content: center;
+  font-size: 12px;
+  color: #fff;
+  font-weight: bold;
+}
+
+/* Post flair badges */
+.post-flair {
+  font-size: 10px;
+  padding: 2px 8px;
+  border-radius: 999px;
+}
+.post-flair.insight { background: rgba(251,191,36,0.15); color: #fbbf24; }
+.post-flair.question { background: rgba(167,139,250,0.15); color: #a78bfa; }
+.post-flair.announcement { background: rgba(239,68,68,0.15); color: #ef4444; }
+.post-flair.bug { background: rgba(239,68,68,0.15); color: #ef4444; }
+.post-flair.idea { background: rgba(74,222,128,0.15); color: #4ade80; }
+.post-flair.meta { background: rgba(136,136,136,0.15); color: #888; }
+
+/* Board detail */
+.board-detail { padding: 20px; }
+.board-detail .post-body { font-size: 14px; line-height: 1.7; margin: 15px 0; }
+
+/* Board comments */
+.board-comment {
+  margin-left: 40px;
+  padding: 10px 12px;
+  background: rgba(255,255,255,0.03);
+  border-radius: 8px;
+  border-left: 2px solid rgba(167,139,250,0.3);
+  margin-bottom: 8px;
+}
+.board-comment .comment-author { font-size: 12px; font-weight: 600; color: #a78bfa; }
+.board-comment .comment-text { font-size: 13px; margin-top: 4px; }
+.board-comment .comment-time { font-size: 11px; color: #666; }
+
+/* ── Activity Tab ──────────────────────────────────────────── */
+.activity-agents {
+  display: flex;
+  gap: 8px;
+  flex-wrap: wrap;
+  margin-bottom: 15px;
+}
+.activity-line {
+  font-family: 'SF Mono', 'Fira Code', monospace;
+  font-size: 13px;
+  padding: 6px 10px;
+  border-left: 2px solid #22d3ee;
+  margin-bottom: 4px;
+}
+
+/* ── Journal ───────────────────────────────────────────────── */
+.journal-list {
+  max-height: 400px;
+  overflow-y: auto;
+  display: flex;
+  flex-direction: column;
+  gap: 4px;
+}
+.journal-entry {
+  display: flex;
+  gap: 10px;
+  padding: 6px 10px;
+  font-size: 13px;
+  font-family: monospace;
+  border-bottom: 1px solid rgba(255,255,255,0.04);
+  align-items: baseline;
+}
+.journal-type {
+  font-size: 11px;
+  min-width: 100px;
+  font-weight: 600;
+}
+.journal-data {
+  flex: 1;
+  color: #aaa;
+  overflow: hidden;
+  text-overflow: ellipsis;
+  white-space: nowrap;
+}
+
