+++ dashboard/src/styles/keeper.css
@@ -1,0 +1,276 @@
+/* ── Keeper Detail Modal ───────────────────────────────────── */
+.keeper-detail-overlay {
+  /* Uses --z-overlay-keeper */
+  position: fixed;
+  top: 0; left: 0; right: 0; bottom: 0;
+  background: rgba(0,0,0,0.6);
+  backdrop-filter: blur(4px);
+  z-index: var(--z-overlay-keeper, 3020);
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  isolation: isolate;
+}
+.keeper-detail {
+  background: #1a1a2e;
+  border: 1px solid rgba(255,255,255,0.12);
+  border-radius: 16px;
+  width: 90vw;
+  max-width: 1100px;
+  max-height: 85vh;
+  overflow-y: auto;
+  overflow-x: visible;
+  padding: 24px;
+  position: relative;
+}
+.keeper-detail-header {
+  display: flex;
+  justify-content: space-between;
+  align-items: flex-start;
+  margin-bottom: 20px;
+}
+.keeper-detail-title { font-size: 22px; font-weight: bold; }
+.keeper-detail-sub { font-size: 13px; color: #888; margin-top: 4px; }
+.keeper-detail-close {
+  background: rgba(255,255,255,0.1);
+  border: none;
+  color: #ccc;
+  font-size: 20px;
+  width: 36px;
+  height: 36px;
+  border-radius: 8px;
+  cursor: pointer;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+.keeper-detail-close:hover { background: rgba(255,255,255,0.2); }
+
+/* Keeper Toolbar */
+.keeper-toolbar {
+  display: flex;
+  gap: 8px;
+  margin-bottom: 16px;
+  flex-wrap: wrap;
+}
+.keeper-toolbar button {
+  padding: 6px 14px;
+  border: 1px solid rgba(255,255,255,0.12);
+  border-radius: 8px;
+  background: rgba(255,255,255,0.05);
+  color: #ccc;
+  cursor: pointer;
+  font-size: 12px;
+}
+.keeper-toolbar button:hover { background: rgba(255,255,255,0.1); }
+.keeper-toolbar button.active {
+  background: rgba(74,222,128,0.15);
+  color: #4ade80;
+  border-color: rgba(74,222,128,0.3);
+}
+
+/* Keeper Detail Grid */
+.keeper-detail-grid {
+  display: grid;
+  grid-template-columns: 1fr 1fr;
+  gap: 16px;
+}
+
+/* Keeper KPIs */
+.keeper-kpis {
+  display: grid;
+  grid-template-columns: repeat(4, 1fr);
+  gap: 10px;
+  margin-bottom: 16px;
+}
+.keeper-kpi {
+  background: rgba(255,255,255,0.04);
+  border-radius: 8px;
+  padding: 12px;
+  text-align: center;
+  border: 1px solid rgba(255,255,255,0.06);
+}
+.keeper-kpi-label { font-size: 10px; text-transform: uppercase; color: #888; letter-spacing: 1px; }
+.keeper-kpi-value { font-size: 22px; font-weight: bold; color: #4ade80; margin-top: 4px; }
+.keeper-kpi-hint { font-size: 10px; color: #666; margin-top: 2px; }
+
+/* Keeper Chart */
+.keeper-chart-card {
+  background: rgba(255,255,255,0.03);
+  border-radius: 10px;
+  padding: 14px;
+  border: 1px solid rgba(255,255,255,0.06);
+}
+.keeper-chart-container { height: 180px; position: relative; }
+.keeper-chart-meta { font-size: 11px; color: #888; margin-top: 8px; }
+.keeper-chart-meta .warn { color: #fbbf24; }
+.keeper-chart-meta .bad { color: #ef4444; }
+
+/* Keeper Field Dictionary */
+.keeper-field-dict {
+  max-height: 460px;
+  overflow-y: auto;
+}
+.keeper-field-search {
+  width: 100%;
+  padding: 8px 12px;
+  background: rgba(255,255,255,0.06);
+  border: 1px solid rgba(255,255,255,0.1);
+  border-radius: 8px;
+  color: #e0e0e0;
+  font-size: 13px;
+  margin-bottom: 10px;
+}
+.keeper-field-search:focus { outline: none; border-color: rgba(74,222,128,0.4); }
+.keeper-field-row {
+  display: flex;
+  gap: 10px;
+  padding: 6px 0;
+  border-bottom: 1px solid rgba(255,255,255,0.04);
+  font-size: 13px;
+}
+.keeper-field-title { font-weight: 600; min-width: 90px; }
+.keeper-field-key { font-family: monospace; color: #22d3ee; font-size: 12px; }
+
+.keeper-signal-list {
+  display: flex;
+  flex-direction: column;
+  gap: 6px;
+}
+
+.keeper-signal-row {
+  display: flex;
+  justify-content: space-between;
+  gap: 10px;
+  padding: 7px 9px;
+  border: 1px solid rgba(138, 163, 211, 0.18);
+  border-radius: 8px;
+  background: rgba(255, 255, 255, 0.03);
+  font-size: 12px;
+}
+
+.keeper-metric-row {
+  align-items: flex-start;
+}
+
+.keeper-metric-meta {
+  display: flex;
+  flex-direction: column;
+  gap: 3px;
+  max-width: 70%;
+}
+
+.keeper-metric-source {
+  color: #8db0dd;
+  font-size: 10px;
+  line-height: 1.25;
+  word-break: break-all;
+}
+
+.keeper-signal-row strong {
+  color: #d9ebff;
+  font-variant-numeric: tabular-nums;
+  font-feature-settings: "tnum";
+}
+
+.keeper-memory-note {
+  margin-top: 10px;
+  border: 1px solid rgba(138, 163, 211, 0.2);
+  border-radius: 8px;
+  background: rgba(255, 255, 255, 0.02);
+  padding: 10px;
+  font-size: 12px;
+  color: #cfe0ff;
+  line-height: 1.5;
+}
+
+/* Keeper Handoff Timeline */
+.keeper-handoff-list {
+  max-height: 220px;
+  overflow-y: auto;
+}
+.keeper-handoff-controls { margin-bottom: 10px; }
+.keeper-handoff-row {
+  border-left: 3px solid #fbbf24;
+  padding: 8px 12px;
+  margin-bottom: 8px;
+  background: rgba(255,255,255,0.03);
+  border-radius: 0 8px 8px 0;
+}
+.keeper-handoff-row summary { cursor: pointer; font-size: 13px; }
+.keeper-handoff-gen { font-size: 11px; color: #22d3ee; }
+.keeper-handoff-time { font-size: 11px; color: #888; }
+.keeper-handoff-meta { font-size: 12px; color: #aaa; margin-top: 6px; }
+.keeper-handoff-trace { font-family: monospace; font-size: 11px; color: #666; }
+
+/* Keeper Events */
+.keeper-events-list {
+  max-height: 180px;
+  overflow-y: auto;
+}
+.keeper-event { padding: 4px 8px; font-size: 12px; border-radius: 4px; margin-bottom: 4px; }
+.keeper-event.handoff { background: rgba(251,191,36,0.08); border-left: 2px solid #fbbf24; }
+.keeper-event.compaction { background: rgba(74,222,128,0.08); border-left: 2px solid #4ade80; }
+
+/* Keeper Equipment */
+.keeper-equipment-list { display: flex; flex-direction: column; gap: 6px; }
+.keeper-equipment-row {
+  display: flex; justify-content: space-between;
+  padding: 6px 10px;
+  background: rgba(255,255,255,0.03);
+  border-radius: 6px;
+  font-size: 13px;
+}
+.keeper-gen-label { color: #22d3ee; font-size: 11px; }
+
+/* Keeper Memory */
+.keeper-memory-list { display: flex; flex-direction: column; gap: 6px; max-height: 220px; overflow-y: auto; }
+.keeper-memory-item {
+  padding: 8px 10px;
+  background: rgba(255,255,255,0.03);
+  border-radius: 6px;
+}
+.keeper-memory-kind { font-size: 10px; text-transform: uppercase; color: #22d3ee; letter-spacing: 1px; }
+.keeper-memory-text { font-size: 13px; margin-top: 4px; }
+.keeper-memory-meta { font-size: 11px; color: #888; margin-top: 2px; }
+
+/* Keeper Conversations */
+.keeper-conversation-list { max-height: 220px; overflow-y: auto; display: flex; flex-direction: column; gap: 6px; }
+.keeper-conversation-item {
+  padding: 8px 10px;
+  background: rgba(255,255,255,0.03);
+  border-radius: 6px;
+}
+.keeper-role-chip {
+  font-size: 10px;
+  padding: 1px 6px;
+  border-radius: 999px;
+  text-transform: uppercase;
+}
+.keeper-role-chip.user { background: rgba(34,211,238,0.15); color: #22d3ee; }
+.keeper-role-chip.assistant { background: rgba(167,139,250,0.15); color: #a78bfa; }
+.keeper-role-chip.warn { background: rgba(251,191,36,0.15); color: #fbbf24; }
+.keeper-role-chip.bad { background: rgba(239,68,68,0.15); color: #ef4444; }
+
+.keeper-conversation-text { font-size: 13px; margin-top: 4px; }
+
+/* Keeper K2K */
+.keeper-k2k-list { max-height: 220px; overflow-y: auto; display: flex; flex-direction: column; gap: 6px; }
+.keeper-k2k-chip {
+  font-size: 10px;
+  padding: 1px 6px;
+  border-radius: 999px;
+  background: rgba(251,191,36,0.15);
+  color: #fbbf24;
+}
+.keeper-k2k-route { font-family: monospace; font-size: 11px; color: #888; }
+
+/* Keeper Mentions */
+.keeper-mention-chip {
+  font-size: 11px;
+  padding: 2px 8px;
+  border-radius: 999px;
+  background: rgba(74,222,128,0.12);
+  color: #4ade80;
+}
+
