+++ dashboard/src/styles/tasks.css
@@ -1,0 +1,50 @@
+/* ── Tasks ─────────────────────────────────────────────────── */
+.task-list { display: flex; flex-direction: column; gap: 6px; }
+.task-row {
+  display: flex;
+  align-items: center;
+  gap: 10px;
+  padding: 8px 12px;
+  background: rgba(255,255,255,0.03);
+  border-radius: 8px;
+  border: 1px solid rgba(255,255,255,0.06);
+}
+.task-info { flex: 1; }
+.task-title { font-size: 14px; }
+.task-assignee { font-size: 12px; color: #888; margin-left: 8px; }
+
+/* Task Board (Kanban) */
+.task-board {
+  display: grid;
+  grid-template-columns: repeat(3, 1fr);
+  gap: 15px;
+}
+.task {
+  padding: 10px 12px;
+  background: rgba(255,255,255,0.04);
+  border-radius: 8px;
+  border-left: 3px solid #888;
+  margin-bottom: 6px;
+}
+.task.todo { border-left-color: #888; }
+.task.in-progress { border-left-color: #fbbf24; }
+.task.done { border-left-color: #4ade80; opacity: 0.7; }
+
+/* ── Messages ──────────────────────────────────────────────── */
+.message-list {
+  max-height: 300px;
+  overflow-y: auto;
+  display: flex;
+  flex-direction: column;
+  gap: 6px;
+}
+.message-row {
+  padding: 8px 12px;
+  background: rgba(255,255,255,0.03);
+  border-radius: 8px;
+  font-size: 13px;
+}
+.message-agent { font-weight: 600; color: #22d3ee; margin-right: 8px; }
+.message-text { color: #ccc; }
+.message-time { font-size: 11px; color: #666; float: right; }
+
